requirements.txt
python-dotenv
rpi.gpio
pigpio
paho-mqtt
